refactor(database): conditionally import mongodb-memory-server
If the proyect is on test mode

+++ database/index.js
@@ -1,3 +1,5 @@
+/* eslint-disable import/no-extraneous-dependencies */
+/* eslint-disable global-require */
 /* eslint-disable no-await-in-loop */
 /* eslint-disable guard-for-in */
 /* eslint-disable no-restricted-syntax */
@@ -5,33 +7,35 @@
 const mongoose = require('mongoose');
 const dotenv = require('dotenv');
 const chalk = require('chalk');
-const { MongoMemoryServer } = require('mongodb-memory-server');
 
 const { log } = console;
-const mongod = new MongoMemoryServer();
 
 dotenv.config();
 mongoose.set('useFindAndModify', false);
 
-module.exports.connectMock = async () => {
-  const URI = await mongod.getUri();
+if (process.env.NODE_ENV === 'test') {
+  const { MongoMemoryServer } = require('mongodb-memory-server');
+  const mongod = new MongoMemoryServer();
+  module.exports.connectMock = async () => {
+    const URI = await mongod.getUri();
 
-  await mongoose
-    .connect(URI, {
-      useNewUrlParser: true,
-      useUnifiedTopology: true,
-    })
-    .then(() => {
-      log(chalk.blue('Connected to MongoDB mock!'));
-    })
-    .catch((err) => log(chalk.red(err)));
-};
+    await mongoose
+      .connect(URI, {
+        useNewUrlParser: true,
+        useUnifiedTopology: true,
+      })
+      .then(() => {
+        log(chalk.blue('Connected to MongoDB mock!'));
+      })
+      .catch((err) => log(chalk.red(err)));
+  };
 
-module.exports.close = async () => {
-  await mongoose.connection.dropDatabase();
-  await mongoose.connection.close();
-  await mongod.stop();
-};
+  module.exports.close = async () => {
+    await mongoose.connection.dropDatabase();
+    await mongoose.connection.close();
+    await mongod.stop();
+  };
+}
 
 module.exports.clear = async () => {
   const { collections } = mongoose.connection;
